# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trends Laravel API",
        "description": "企業の技術コミュニティでの影響力を分析・追跡するAPI",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/companies/{company_id}": {
            "get": {
                "tags": [
                    "企業詳細"
                ],
                "summary": "企業詳細情報取得",
                "description": "企業の詳細情報を取得します。現在のランキング情報と最近の記事（最大5件）を含みます。",
                "operationId": "b87956b293dcfbb97d69bf77d6d496cc",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "企業ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "企業詳細情報",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Company A"
                                                },
                                                "domain": {
                                                    "type": "string",
                                                    "example": "company-a.com"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "企業の説明文"
                                                },
                                                "logo_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/logo.png"
                                                },
                                                "website_url": {
                                                    "type": "string",
                                                    "example": "https://company-a.com"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "current_rankings": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "recent_articles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "企業IDが無効です"
                                        },
                                        "details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "企業が見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "企業が見つかりません"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/{company_id}/articles": {
            "get": {
                "tags": [
                    "企業詳細"
                ],
                "summary": "企業の記事一覧取得",
                "description": "企業に関連する記事の一覧を取得します。",
                "operationId": "a6fca0b29fa759399957964775c3bb6f",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "企業ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "ページ番号",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "1ページあたりの件数（最大100）",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "example": 20
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "過去何日分の記事を取得するか",
                        "schema": {
                            "type": "integer",
                            "default": 30,
                            "example": 30
                        }
                    },
                    {
                        "name": "min_bookmarks",
                        "in": "query",
                        "description": "最小ブックマーク数",
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "記事一覧",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "company_id": {
                                                    "type": "integer"
                                                },
                                                "filters": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "企業IDが無効です"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "企業が見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "企業が見つかりません"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}